src/htbla/aud3/graphtheory/Edge.java
package htbla.aud3.graphtheory;

import java.util.Objects;

/*
 * @author Jonas Heschl, Jakob Jodlbauer, Emil Maximilian Hoani Eichsteininger
 */
public class Edge {

    private int from;
    private int to;
    private int weight;

    public Edge(int from, int to, int weight) {
        this.from = from;
        this.to = to;
        this.weight = weight;
    }

    public int getFirstNodeId() {
        return from;
    }
    
    public int getSecondNodeId() {
        return to;
    }

    public int getWeight() {
        return weight;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Edge edge = (Edge) o;
        return from == edge.from && to == edge.to && weight == edge.weight;
    }

    @Override
    public int hashCode() {
        return Objects.hash(from, to, weight);
    }

    @Override
    public String toString() {
        return "Edge{" +
                "from=" + from +
                ", to=" + to +
                ", weight=" + weight +
                '}';
    }
}
